Add files via upload
This is my final design for the type specimen poster. I chose to present the poster as a ‘ticket’ as this font is universally used for materials such as tickets, books and other everyday publications because it is easy to read. I chose to add the information about the font in a ‘barcode’ shape at the bottom right of the ticket as I knew it would be more efficient to disguise text as a barcode. 
On the poster I have also added pieces of history about the font, for example the year it was created and the creator of the font. I also put the price as euro as the font originates from Germany. The red background colour was chosen due to the Germany flag containing red and the font is black which also ties the flag into the poster. I didn’t want to make it blatantly obvious that I was connecting the flag into the poster so that is why I used white in the background appose to yellow. This is not only more visually appealing but a more subtle approach to get my point across.